more checks

+++ cut_off.py
@@ -131,7 +131,9 @@ def fetch():
         os.rename((BUILD_O5M), (BUILD_OLD))
         
       os.rename((BUILD_TEMP), (BUILD_O5M))
-      os.remove(BUILD_OLD)
+      ExitCode = os.path.exists(BUILD_OLD) 
+      if ExitCode == True:
+        os.remove(BUILD_OLD)
       
     else:
       printerror("tiles/" + (BUILDMAP) + ".poly not found... exit")
